backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.2.0
pdf2image==1.17.0

# OCR & Computer Vision
pytesseract==0.3.10
opencv-python==4.9.0.80

# Video Processing
moviepy==1.0.3
SpeechRecognition==3.10.1
pydub==0.25.1

# AI/ML
openai==1.10.0
tiktoken==0.5.2

# Background Processing
rq==1.16.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-magic-bin==0.4.14
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
diskcache==5.6.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.24.0
sse-starlette==1.8.2
langdetect==1.0.9
openpyxl==3.1.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.14
